+++ CheckIt/CheckIt/Extensions/Ex+Date.swift
@@ -22,4 +22,24 @@ extension Date {
         
         return dateFormatter.string(from: self)
     }
+    
+    func yearMonthDayDateToString(date: Date) -> String {
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "yyyy년 MM월 dd일"
+        let dateString = dateFormatter.string(from: date)
+        return dateString
+    }
+    func monthDayDateToString(date: Date) -> String {
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "MM월 dd일"
+        let dateString = dateFormatter.string(from: date)
+        return dateString
+    }
+    func hourMinuteDateToString(date: Date) -> String {
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "a h시 mm분"
+        dateFormatter.locale = Locale(identifier: "ko_KR")
+        let dateString = dateFormatter.string(from: date)
+        return dateString
+    }
 }
